Refactor Plans component improved modularity and maintainability
- Extracted , , and  into separate components.
- Updated  component to use the new modular components.
- Refactored  to include reusable dropdown components for frequency and duration.
- Updated  and  to handle rendering tasks and goals dynamically.
- Improved styling and structure consistency across components using Tailwind CSS.
